fix(media): relocate hydrate_part/hydrate_prompt_parts out of primer.channel
primer/agent/loop.py imported primer.channel.media for prompt-time
hydration, tripping the "agent/worker reach channel only through
sanctioned dispatch seams" import-linter contract (CI-only step, not
caught locally). The two functions are layer-pure -- they only touch
Message/Part and the ArtifactStorage interface -- so they belong below
primer.channel in the dependency graph, not behind it. Moved them to
the new primer.media.hydrate module; primer.channel.media re-exports
both names for its existing callers (adapters, tests).
