+++ primer/channel/media.py
@@ -7,8 +7,12 @@ No platform SDKs here. This module:
 * enforces size + type limits,
 * stores inbound bytes as an artifact and builds the referencing part,
 * collects media parts from a turn's output (outbound), and
-* hydrates an artifact-referencing part back to inline ``data`` (for the LLM
-  at turn time and for a channel upload at relay time).
+* parses ``PromptEnvelope.media`` dicts and hydrates them for channel upload.
+
+``hydrate_part``/``hydrate_prompt_parts`` (resolving an artifact-referencing
+part back to inline ``data``) live in :mod:`primer.media.hydrate` and are
+re-exported here for backward compatibility -- the agent turn loop needs
+them too and may not import ``primer.channel``.
 
 Bytes live in the :class:`primer.int.artifact_storage.ArtifactStorage` backend;
 persisted parts carry only an ``artifact_id`` reference.
@@ -21,6 +25,7 @@ import logging
 from dataclasses import dataclass, field
 
 from primer.int.artifact_storage import ArtifactStorage
+from primer.media.hydrate import hydrate_part, hydrate_prompt_parts
 from primer.model.chat import (
     AudioPart, DocumentPart, ImagePart, Message, Part, VideoPart,
 )
@@ -269,52 +274,6 @@ async def hydrate_media_dicts(artifact_registry, media: list[dict] | None) -> li
             out.append(await hydrate_part(store, part))
         except Exception:
             logger.warning("hydrate_media_dicts: hydrate failed; skipping")
-    return out
-
-
-async def hydrate_part(artifact_storage: ArtifactStorage, part: Part) -> Part:
-    """Return a copy of ``part`` with inline ``data`` populated from its
-    ``artifact_id`` (when set and ``data`` is empty). Clears ``artifact_id`` on
-    the copy so downstream consumers see only ``data``. Parts without an
-    ``artifact_id`` (or that already carry ``data``) pass through unchanged."""
-    aid = getattr(part, "artifact_id", None)
-    if not aid or getattr(part, "data", None):
-        return part
-    blob = await artifact_storage.get(aid)
-    if blob is None:
-        logger.warning("hydrate_part: artifact %s not found", aid)
-        return part
-    update: dict = {"data": blob.data, "artifact_id": None}
-    if getattr(part, "mime_type", None) is None:
-        update["mime_type"] = blob.mime_type
-    return part.model_copy(update=update)
-
-
-async def hydrate_prompt_parts(
-    artifact_storage: ArtifactStorage | None, messages: list[Message],
-) -> list[Message]:
-    """Resolve every part's ``artifact_id`` to inline ``data`` across a
-    whole prompt, so an LLM adapter (which only reads ``data``/``url``/
-    ``file_id``, never ``artifact_id``) sees real bytes for any
-    artifact-backed part in history or the turn's new messages.
-
-    ``artifact_storage=None`` returns ``messages`` unchanged (no-op) --
-    callers that never resolve a store (tests, surfaces with no
-    ArtifactStorage wired) keep today's behaviour exactly. Messages with
-    no binary parts are returned as the SAME object (no copy), so this is
-    cheap to call unconditionally on every turn.
-    """
-    if artifact_storage is None:
-        return messages
-    out: list[Message] = []
-    for msg in messages:
-        if not any(getattr(p, "artifact_id", None) for p in msg.parts):
-            out.append(msg)
-            continue
-        hydrated_parts = [
-            await hydrate_part(artifact_storage, p) for p in msg.parts
-        ]
-        out.append(msg.model_copy(update={"parts": hydrated_parts}))
     return out
